requirements.txt
streamlit
openai
python-dotenv
python-docx
PyMuPDF
requests